+++ source/javascripts/components/unified-editor/TriggersContent/RegexCheckbox.tsx
